# api-lib/openapi.yml
openapi: 3.1.0
info:
  title: Wish site OpenAPI spec
  version: 0.0.2
servers:
  - url: http://localhost:8080/api/
    description: Dev Server
tags:
  - name: Users
    description: Управление пользователями
  - name: Wishes
    description: Управление желаниями
  - name: Auth
    description: Авторизация и получение токенов
  - name: Files
    description: Управление файлами (загрузка, удаление)

paths:
  # --- USERS ---
  /users:
    get:
      summary: Метод получения всех юзеров
      tags:
        - Users
      operationId: getAllUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Успешный ответ со списком всех юзеров
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        '400':
          description: Ошибка
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      summary: Метод создания юзера
      tags:
        - Users
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequest"
      responses:
        '200':
          description: Успешный ответ с созданным юзером
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        description: Идентификатор юзера
        schema:
          type: integer
          format: int64
    get:
      security:
        - bearerAuth: []
      summary: Метод получения конкретного user по id
      tags:
        - Users
      operationId: getUserById
      responses:
        '200':
          description: Успешный ответ с конкретным юзером
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        '404':
          description: Юзер не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      security:
        - bearerAuth: []
      summary: Метод удаления юзера по id
      tags:
        - Users
      operationId: deleteUserById
      responses:
        '200':
          description: "Успешный ответ: юзер с user_id удален"
        '404':
          description: Юзер не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    patch:
      security:
        - bearerAuth: []
      summary: Обновление данных юзера по id
      tags:
        - Users
      operationId: updateUserById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        '200':
          description: Успешный ответ с обновленным юзером
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        '400':
          description: Ошибка валидации

  # --- WISHES ---


  /users/{user_id}/wishes:
    parameters:
      - name: user_id
        in: path
        required: true
        description: ID владельца желаний
        schema:
          type: integer
          format: int64

    get:
      security:
        - bearerAuth: []
      summary: Получить все желания конкретного пользователя
      tags:
        - Wishes
      operationId: getUserWishes
      parameters:
        - name: page
          in: query
          required: false
          description: Номер страницы (0-based)
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Количество элементов на странице
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Пагинированный список желаний пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishPageResponse"

    post:
      security:
        - bearerAuth: []
      summary: Создать новое желание для пользователя
      tags:
        - Wishes
      operationId: createWish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishRequest"
      responses:
        '200':
          description: Желание успешно создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishResponse"

  /users/{user_id}/wishes/with-file:
    post:
      security:
        - bearerAuth: []
      summary: Создать новое желание с файлом (транзакционно)
      description: |
        Создает wish с файлом в одном запросе. Использует двухфазный коммит для распределенной транзакции между БД и MinIO.
        При ошибке автоматически откатывает изменения в обоих системах.
      tags:
        - Wishes
      operationId: createWishWithFile
      parameters:
        - name: user_id
          in: path
          required: true
          description: Идентификатор пользователя
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateWishWithFileRequest"
      responses:
        '200':
          description: Желание успешно создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishResponse"
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Ошибка при создании wish или загрузке файла
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /wishes/{wish_id}:
    parameters:
      - name: wish_id
        in: path
        required: true
        description: ID желания
        schema:
          type: integer
          format: int64

    get:
      security:
        - bearerAuth: []
      summary: Получить детали конкретного желания
      tags:
        - Wishes
      operationId: getWishById
      responses:
        '200':
          description: Детали желания
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishResponse"
        '404':
          description: Желание не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      security:
        - bearerAuth: []
      summary: Удалить желание
      tags:
        - Wishes
      operationId: deleteWishById
      responses:
        '200':
          description: Желание удалено
        '404':
          description: Желание не найдено

    patch:
      security:
        - bearerAuth: []
      summary: Обновить желание
      tags:
        - Wishes
      operationId: updateWishById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishUpdateRequest"
      responses:
        '200':
          description: Желание обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishResponse"

  /wishes/{wish_id}/with-file:
    patch:
      security:
        - bearerAuth: []
      summary: Обновить желание с файлом (транзакционно)
      description: |
        Обновляет wish с файлом в одном запросе. Использует двухфазный коммит для распределенной транзакции между БД и MinIO.
        При ошибке автоматически откатывает изменения в обоих системах.
      tags:
        - Wishes
      operationId: updateWishByIdWithFile
      parameters:
        - name: wish_id
          in: path
          required: true
          description: ID желания
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateWishWithFileRequest"
      responses:
        '200':
          description: Желание успешно обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishResponse"
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Желание не найдено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Ошибка при обновлении wish или загрузке файла
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- AUTH ---
  /auth/login:
    post:
      security: []
      summary: Авторизация пользователя
      description: Принимает логин и пароль и возвращает access_token
      tags:
        - Auth
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: Авторизация прошла успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '429':
          description: Слишком много попыток входа
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /auth/register:
    post:
      security: []
      summary: Регистрация нового пользователя
      description: Создает нового пользователя и возвращает access_token
      tags:
        - Auth
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        '200':
          description: Регистрация прошла успешно
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '400':
          description: Ошибка валидации или пользователь уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- FILES ---
  /files/upload:
    post:
      summary: Загрузить файл (фото) в MinIO
      description: Загружает изображение в MinIO и возвращает URL для доступа к файлу. Максимальный размер - 10MB. Поддерживаются только изображения.
      tags:
        - Files
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Файл изображения (jpg, png, gif, webp)
      responses:
        '200':
          description: Файл успешно загружен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploadResponse"
        '400':
          description: Ошибка валидации (файл не предоставлен, неверный формат или слишком большой)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Ошибка при загрузке файла в MinIO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /files/{object_name}:
    delete:
      summary: Удалить файл из MinIO
      description: Удаляет файл из MinIO по его object name (например, wishes/uuid.jpg)
      tags:
        - Files
      operationId: deleteFile
      security:
        - bearerAuth: []
      parameters:
        - name: object_name
          in: path
          required: true
          description: Object name файла в MinIO (например, wishes/550e8400-e29b-41d4-a716-446655440000.jpg)
          schema:
            type: string
          example: "wishes/550e8400-e29b-41d4-a716-446655440000.jpg"
      responses:
        '200':
          description: Файл успешно удален
        '400':
          description: Некорректный object name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    # --- USER MODELS ---
    UserResponse:
      type: object
      required:
        - user_id
        - name
        - role
        - created_at
      properties:
        user_id:
          type: integer
          format: int64
          example: 1337
        name:
          type: string
          example: Artem Dabstep
        role:
          type: string
          description: "Роль юзера: Админ или обычный"
          enum:
            - admin
            - member
        created_at:
          type: string
          format: date-time
          example: "2025-11-27T16:44:00+03:00"

    UserRequest:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
          example: Artem Dabstep
        password:
          type: string
          format: password
          example: "SuperSecret"

    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: New Name 123
          description: Новое имя (опционально)
        role:
          type: string
          enum:
            - admin
            - member
          default: member
          description: Новая роль (опционально)
        password:
          type: string
          format: password
          description: Новый пароль (опционально)

    # --- WISH MODELS ---
    WishResponse:
      type: object
      required:
        - wish_id
        - title
        - created_at
        - owner_id
      properties:
        wish_id:
          type: integer
          format: int64
          example: 101
        owner_id:
          type: integer
          format: int64
          description: ID пользователя, создавшего желание
          example: 1337
        title:
          type: string
          description: Название желания
          example: "PlayStation 5"
        description:
          type: string
          description: Подробное описание
          example: "С дисководом, желательно белую"
        photo_url:
          type: string
          description: Ссылка на фото (из MinIO)
          example: "https://minio.dabwish.com/images/ps5.jpg"
        price:
          type: number
          format: double
          description: Ориентировочная цена
          example: 499.99
        created_at:
          type: string
          format: date-time
          example: "2025-12-01T10:00:00+03:00"

    WishPageResponse:
      type: object
      required:
        - items
        - page
        - size
        - total_elements
        - total_pages
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WishResponse"
        page:
          type: integer
          format: int32
          description: Текущий номер страницы (0-based)
        size:
          type: integer
          format: int32
          description: Размер страницы
        total_elements:
          type: integer
          format: int64
          description: Общее количество желаний
        total_pages:
          type: integer
          format: int32
          description: Общее количество страниц

    WishRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: "PlayStation 5"
        description:
          type: string
          example: "С дисководом"
        photo_url:
          type: string
          description: Ссылка на уже загруженное фото
        price:
          type: number
          format: double

    CreateWishWithFileRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: "PlayStation 5"
          description: Название желания
        description:
          type: string
          example: "С дисководом"
          description: Описание желания
        photo:
          type: string
          format: binary
          description: Файл изображения (опционально, максимум 10MB, только изображения)
        price:
          type: number
          format: double
          description: Ориентировочная цена
          example: 49999.99

    WishUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        photo_url:
          type: string
        price:
          type: number
          format: double

    UpdateWishWithFileRequest:
      type: object
      properties:
        title:
          type: string
          description: Название желания
          example: "PlayStation 5"
        description:
          type: string
          description: Описание желания
          example: "С дисководом"
        photo:
          type: string
          format: binary
          description: Файл изображения (опционально, максимум 10MB, только изображения)
        price:
          type: number
          format: double
          description: Ориентировочная цена
          example: 49999.99

    # --- AUTH MODELS ---
    LoginRequest:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
          description: Логин пользователя
          example: admin
        password:
          type: string
          format: password
          description: Пароль пользователя
          example: qwerty123

    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: JWT access token
          example: fjdkfKFSKLFDjfdkfjkd...
        token_type:
          type: string
          description: Тип токена
          example: Bearer
          default: Bearer
        expires_in:
          type: integer
          format: int64
          description: Время жизни access токена в секундах
          example: 3600
        user:
          $ref: "#/components/schemas/UserResponse"

    RegisterRequest:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
          description: Имя пользователя
          example: new_user
        password:
          type: string
          format: password
          minLength: 6
          description: Пароль пользователя
          example: SuperSecret123

    FileUploadResponse:
      type: object
      required:
        - file_url
        - object_name
      properties:
        file_url:
          type: string
          description: Полный URL для доступа к загруженному файлу
          example: "http://localhost:9000/dabwish-images/wishes/550e8400-e29b-41d4-a716-446655440000.jpg"
        object_name:
          type: string
          description: Имя объекта в MinIO bucket (можно использовать для удаления)
          example: "wishes/550e8400-e29b-41d4-a716-446655440000.jpg"

    # --- COMMON ---
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          example: Resource not found

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT